+++ package.json
@@ -15,7 +15,7 @@
     "db-migrate:test": "NODE_ENV=test ./node_modules/.bin/babel-node ./node_modules/.bin/sequelize db:migrate",
     "db-migrate:undo": "./node_modules/.bin/babel-node ./node_modules/.bin/sequelize db:migrate:undo:all",
     "db-migrate:undo:test": "NODE_ENV=test ./node_modules/.bin/babel-node ./node_modules/.bin/sequelize db:migrate:undo:all",
-    "db-refresh": "npm run db-migrate:undo && npm run db-migrate && npm run db-seed",
+    "db-refresh": "npm run db-drop && npm run db-create && npm run db-migrate:undo && npm run db-migrate && npm run db-seed",
     "db-refresh:test": "NODE_ENV=test npm run db-migrate:undo:test && npm run db-migrate:test && npm run db-seed:test",
     "db-seed": "./node_modules/.bin/babel-node ./node_modules/.bin/sequelize db:seed:all",
     "db-seed:test": "NODE_ENV=test ./node_modules/.bin/babel-node ./node_modules/.bin/sequelize db:seed:all",
